requirements.txt
autopep8==1.5.3
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.4
pycodestyle==2.6.0
toml==0.10.1
xlwt==1.3.0
cassandra-driver==3.24.0
